Move functionality into services, remove unused methods
