+++ src/test/java/utils/TestContext.java
@@ -147,4 +147,10 @@ public class TestContext {
     public FormSendCalloutHandler formSendCalloutHandler() {
         return Mockito.mock(FormplayerFormSendCalloutHandler.class);
     }
+
+    @Bean
+    public MenuSessionRunnerService menuSessionRunnerService() {return Mockito.spy(MenuSessionRunnerService.class);}
+
+    @Bean
+    public MenuSessionFactory menuSessionFactory() {return Mockito.spy(MenuSessionFactory.class);}
 }
